Implemented an important part of the spec, "Attributes omitted from the resource object should not be updated." This requires us to be able to communicate to the IMaterializer (which does actual updating) that attributes were or were not specified on the original JSON resource object. Therefore, we created the new MetadataManager singleton, and its public method PropertyWasPresent(object deserialized, PropertyInfo prop). It is tightly coupled with JsonApiFormatter, but loosely with an IMaterializer such as EntityFrameworkMaterializer.
The ability to check for the presence of a specified value in the JSON also makes it possible to use default values on a POSTed object--before we could not tell if there was underposting going on!

This will bump the version to 0.2a, because the behavior in an underposting scenario has changed.