allow specifying model manager to fluent config
This change lets the user specify a model manager for the fluent config
to use. This means that type registrations will be done directly against
the model manager instead of the fluent config.

This isn't the nicest solution, but without taking over the
IDependencyResolver, I couldn't think of anything better.